separated all potential to its own class
