+++ SourceCode/MDBase.py
@@ -1,21 +1,24 @@
-import functools
 import numpy as np
+import functools
 from ase.io.trajectory import Trajectory
 from ase.md import MDLogger
-from ase.md.velocitydistribution import MaxwellBoltzmannDistribution
+from ase.md.velocitydistribution import MaxwellBoltzmannDistribution, Stationary, ZeroRotation
 from ase.units import fs, GPa
 from ase.visualize import view
 import logging
-from LJRegistry import LJParams, calcMaxRc
 from simulationInput import SimulationSettings
+from potentialSetUp import Potential
 
 log = logging.getLogger(__name__)
+
+
 class MDBase:
     """
     basic MD class
     When initialized it represents a MD-simulation with prefilled settings that can be used for multiple runs with
     different atomic structures.
     """
+
     def __init__(self, settings: SimulationSettings):
         """
         Create a MD runner from the settings specified in the 
@@ -25,15 +28,14 @@ class MDBase:
         
         """
 
-        #General parameters
+        # General parameters
         self.ensemble = settings.ensemble
         self.timestep = settings.timestep * fs
         self.steps = settings.num_steps
         self.interval = settings.sample_interval
         self.output_file = settings.output_file
-        
 
-        #Ensemble specific parameters
+        # Ensemble specific parameters
         if settings.ensemble == "NVE":
             self.temperature_k = settings.initial_temperature
 
@@ -43,87 +45,12 @@ class MDBase:
 
         elif settings.ensemble == "NPT":
             self.temperature_k = settings.temperature
-            self.pressure = settings.pressure * GPa * 1e-9 # Pa to Au
-            self.compressibility = settings.compressibility/(GPa*1e-9) # Pa^-1 to Au
-        
-        #Integrator and potential
+            self.pressure = settings.pressure * GPa * 1e-9  # Pa to Au
+            self.compressibility = settings.compressibility / (GPa * 1e-9)  # Pa^-1 to Au
+
+        # Integrator and potential
         self.integrator = self.getIntegrator(self.ensemble)
-        self.potential = self.getPotential(settings.potential)
-
-
-
-    def setupLJCalculator(self, atoms):
-        symbols = atoms.get_chemical_symbols()
-        uniq = sorted(set(symbols))
-        if len(uniq) != 1:
-            raise ValueError(
-                f"ASE LennardJones supports a single atom type only; found {uniq}. "
-            )
-
-        material_key = uniq[0].lower()
-        params = LJParams(material=material_key)
-        atomic_number = [(atoms.get_atomic_numbers()[0])]
-        eps = params["epsilon_eV"]
-        sig = params["sigma_A"]
-        ro = params["ro_A"]
-        rc = params["rc_A"]
-        rc_max = calcMaxRc(atoms)
-
-        if rc > rc_max:
-            log.warning("The rCut is larger than the cell size, will use cell size to derive new value for rCut")
-            rc = rc_max
-            ro = 0.9*float(rc)
-
-        print("using this rc ", rc)
-        try:
-            from asap3 import LennardJones as asap_LJ
-            calc_asap = asap_LJ(
-                atomic_number,
-                epsilon=[eps],
-                sigma=[sig],
-                rCut=rc,
-                modified=True
-            )
-
-            atoms.calc = calc_asap
-            _ = atoms.get_potential_energy()
-            log.info("Using asap3 LJ | element=%s (Z=%s) | ε=%.4g eV | σ=%.4g Å | rc=%.4g Å ",
-                     material_key, atomic_number[0], eps, sig, rc)
-            return calc_asap
-
-
-        except Exception as e:
-            from ase.calculators.lj import LennardJones as ase_LJ
-            calc_ase = ase_LJ(
-            epsilon=eps,
-            sigma=sig,
-            rc=rc,
-            ro=ro
-            )
-            log.warning(
-                f"Falling back to ASE LJ | Reason: {e}"
-            )
-            return calc_ase
-
-    def getPotential(self, potential: str):
-        """
-        In:
-            String: potential
-        Out:
-            Potential_function: potential
-        """
-        potential_lower = potential.lower()
-        if potential_lower in ["emt"]:
-            from asap3 import EMT as asap_EMT
-            log.info("Potential: EMT")
-            return lambda atoms: asap_EMT()
-
-        elif potential_lower in ["lj", "lennardjones", "lennard_jones"]:
-            log.info("Potential: Lennard Jones")
-            return self.setupLJCalculator
-        else:
-            log.error("Invalid potential function: %s", potential)
-            raise ValueError(f"Invalid potential function: {potential}")
+        self.potential = Potential().getPotential(settings.potential)
 
     def getIntegrator(self, integrator: str):
         """
@@ -162,9 +89,9 @@ class MDBase:
         equilibrium_steps = 20000
         dyn_eq = self.integrator(atoms=atoms)
 
-         #dyn_eq.attach(lambda: self.checkConvergence(atoms), interval=max(1, int(1 / self.timestep)))
-        #log.info(
-          #  f"Starting equilibrium run with {self.ensemble} Ensemble to reach desired temperature of {self.temperature_k} K")
+        dyn_eq.attach(lambda: self.checkConvergence(atoms), interval=max(1, int(1 / self.timestep)))
+        # log.info(
+        #  f"Starting equilibrium run with {self.ensemble} Ensemble to reach desired temperature of {self.temperature_k} K")
 
         try:
             dyn_eq.run(int(equilibrium_steps))
@@ -174,7 +101,7 @@ class MDBase:
         except RuntimeError as e:
             # pep 437 problem
             if "generator raised StopIteration" in str(e):
-                log.info(
+                log.warning(
                     f"Equilibrium reached early (observer signaled StopIteration) at T = {round(atoms.get_temperature(), 2)}.")
             else:
                 raise RuntimeError(e)
@@ -183,7 +110,8 @@ class MDBase:
             log.info(f"Equilibrium reached early: {ok}")
 
         except RuntimeWarning as err:
-            log.warning(f"Equilibrium aborted due to instability: {err}")
+            log.error(f"Equilibrium aborted due to instability: {err}")
+            quit()
 
         finally:
             self.quantity_list = []
@@ -202,11 +130,15 @@ class MDBase:
         MaxwellBoltzmannDistribution(atoms, temperature_K=self.temperature_k,
                                      force_temp=True)  # Initialize velocity according to temperature_k
 
+        # might be better to place these after eq run?
+        Stationary(atoms, preserve_temperature=True)
+        ZeroRotation(atoms, preserve_temperature=True)
+
         self.equilibriumRun(atoms=atoms)
         log.info("MD run starts with: %i steps", self.steps)
         dyn = self.integrator(atoms=atoms)
 
-        traj = Trajectory(filename=f"../{self.output_file}.traj", mode="w", atoms=atoms) ## currently have .. before
+        traj = Trajectory(filename=f"../{self.output_file}.traj", mode="w", atoms=atoms)  ## currently have .. before
 
         # Custom calculation function
         def save_custom_data():
@@ -217,14 +149,14 @@ class MDBase:
 
         dyn.attach(save_custom_data, interval=self.interval)
 
-        #for a in self.attachments:
-         #   dyn.attach(functools.partial(a, atoms=atoms),
-         #              interval=self.interval)  # Attach the different functions for printing
+        # for a in self.attachments:
+        #   dyn.attach(functools.partial(a, atoms=atoms),
+        #              interval=self.interval)  # Attach the different functions for printing
 
         dyn.attach(traj.write, interval=self.interval)
 
         dyn.attach(lambda: self.checkDivergence(atoms),
-                   interval=self.interval)  # TODO Possibly include checkConvergence here?
+                   interval=self.interval)
 
         # Apply a short sequence of slight, controlled strains and run a few steps at each.
         # This creates trajectory frames with non-zero strain for robust post-processing of elastic constants.
@@ -232,7 +164,7 @@ class MDBase:
             A = atoms.cell.array.T
             A_new = (F @ A).T
             atoms.set_cell(A_new, scale_atoms=True)
-            #log.info(f"Applied strain F=\n{F}\nCell now: {atoms.cell.cellpar()}")
+            # log.info(f"Applied strain F=\n{F}\nCell now: {atoms.cell.cellpar()}")
             dyn.run(int(steps))
 
         # Small strain amplitude
@@ -260,7 +192,7 @@ class MDBase:
         eps_yz[1, 2] = eps_yz[2, 1] = eta;
         F_list.append(eps_yz)
 
-        #log.info(f"Starting pre-production strain sequence with {len(F_list)} strains; {hold_steps} steps each")
+        # log.info(f"Starting pre-production strain sequence with {len(F_list)} strains; {hold_steps} steps each")
         for F in F_list:
             _apply_F_and_run(F, hold_steps)
 
@@ -268,87 +200,44 @@ class MDBase:
         dyn.run(self.steps)  # RUN
         traj.close()  # Explicitly close the trajectory
 
-
-
     def checkConvergence(self, atoms):
+        """
+
+        """
         self._updateQuantityList(atoms)
-        if len(self.quantity_list) > 150 and not self.checkInstability(self.quantity_list):
-            match self.ensemble:
-                case "NVE":
-                        raise StopIteration(f"Converged: E = {(atoms.get_potential_energy() + atoms.get_kinetic_energy()) / len(atoms):.2f} ev/atom")
-                case "NVT":
-                        raise StopIteration(f"Converged: T = {atoms.get_temperature():.2f} K")
-                case "NPT":
-                        raise StopIteration(f"Converged: T = {atoms.get_temperature():.2f} K")
+        window = 500
+        error = 1e-6
+        if len(self.quantity_list) < window:
+            return
 
-
-    def checkDivergence(self, atoms):
-        """
-        Checks if temperature or energy diverges continously in one direction,
-        returns an error if that's the case, uses a 'window' number of last runs to calculate mean average
-        hits determines how many flucuations before exiting
-        """
-        # Uses helper function updateQuantityList.
-        self._updateQuantityList(atoms)
-        if (self.checkInstability(self.quantity_list)):
-            # End the run and raise warning.
-            raise RuntimeError(f"{self.ensemble}: Run cancelled.")
-
-    def checkInstability(self, data_buffer):
-        """
-        Uses an input list to calculate stability, returns false if stable and true if unstable.
-
-        Inputs: List with atomic temperature, energy or volume
-        Output: Bool
-        """
-        window = 10
-        tol_energy_percentage = 0.05
-        tol_temp = self.temperature_k * 0.10
-        nb_outside_tolerance = 0
-        threshold = 100000000 # needs to be large number
-
-        if len(data_buffer) <= 1:
-            return False
-
+        last = np.asarray(self.quantity_list[-window:], dtype=float)
+        mean_diff = abs(last[int(window/2):].mean() - last[:int(window/2)].mean())
+        last_diff = abs(self.quantity_list[-1] - self.quantity_list[-2])
         match self.ensemble:
-            case "NVT" | "NPT":
-                lastN = data_buffer[-window:]
-                nb_outside_tolerance = sum(abs(x - self.temperature_k) >= tol_temp for x in lastN)
-            case "NVE":
-                dt_eff_ps = self.interval * self.timestep / (1000 * fs)
-                num_points = max(3, int(round(3.0 / dt_eff_ps))) # These two integers can be altered
-                if len(data_buffer) < num_points:
-                    return False
-                lastN = np.asarray(data_buffer[-num_points:], dtype=float)
-                mean_energy = float(np.mean(lastN))
-                energy_tol = max(tol_energy_percentage * abs(mean_energy), 1e-4)
-                nb_outside_tolerance = int(np.sum(np.abs(lastN - mean_energy) >= energy_tol))
-                threshold = max(3, int(np.ceil(0.3 * num_points))) # 3 points must be outside of tolerance to trigger
-                # Compares the number of points that are outside the threshold and returns True if they are
+            case "NVT":
+                error_per_atom = error*len(atoms)
+                if last_diff <= error_per_atom and mean_diff <= error_per_atom:
+                        raise StopIteration(f"Converged at E_pot = {last[-1]} eV.")
 
-        if (nb_outside_tolerance >= threshold):
-            log.warning(
-                f"Energy oscillates heavily — {nb_outside_tolerance}/{num_points} points outside tolerance"
-            )
-            return False
-        if 12 < nb_outside_tolerance < 15:
-            # Print warning in log, continues the run
-            log.warning("Warning temperature oscillates.")
-        if (nb_outside_tolerance >= 15):
-            log.warning("Temperature oscillates heavily, cancelling run.")
-            return True
-        #else:
-        #    return (not nb_outside_tolerance > 10)
-        #return False
+            case "NVE":
+                pass
+
+            case "NPT":
+                raise StopIteration(f"Converged: T = {atoms.get_temperature():.2f} K")
+
+
+
 
     def _updateQuantityList(self, atoms):
         """
-        Help function that appends quantities into quantity_list.
+        Help function that appends quantities depending on the ensemble into quantity_list.
         """
         if not hasattr(self, "quantity_list"):
             self.quantity_list = []
         match self.ensemble:
+            case "NPT":
+                self.quantity_list.append(atoms.get_volume())
+            case "NVT":
+                self.quantity_list.append(atoms.get_potential_energy())
             case "NVE":
-                self.quantity_list.append((atoms.get_potential_energy() + atoms.get_kinetic_energy()) / len(atoms))
-            case "NVT" | "NPT":
                 self.quantity_list.append(atoms.get_temperature())
